RavenDB-26643 - move to WebSocketHelper in Raven.Server.Utils
