+++ src/Raven.Server/NotificationCenter/Handlers/ClusterDashboardConnection.cs
@@ -7,6 +7,7 @@ using System.Threading.Tasks;
 using Raven.Server.Dashboard;
 using Raven.Server.Dashboard.Cluster;
 using Raven.Server.Dashboard.Cluster.Notifications;
+using Raven.Server.Utils;
 using Sparrow.Json;
 using Sparrow.Json.Parsing;
 using Sparrow.Logging;
